♻️ Refactor ProfileForm Ui module
