+++ client/components/modules/ProfileForm/ProfileForm.interface.ts
@@ -1,0 +1,7 @@
+
+import { type UpdateProfileDto } from '@/api/profile.api';
+
+export interface ProfileFormProps {
+  initialData?: UpdateProfileDto;
+  onSuccess?: () => void;
+}
